Replaces images with a 2x version
On retina computers, the normal images look awful. Because of this all images (all two...) have been replaced with an image twice the size. This also replaces the old method for making images black/white, with a new sprite method that has better browser support (Firefox!).

Later on, only retina users should get the 2x version, so that the page load is only heavier for the ones that need it.